[RDSS-1900] Create collection, both root and child
Deliberately using `lxml` rather than the build-in `xml`, since the
built-in `xml` library, does slightly unexpected things with namespaces:

  - Changing prefixes

  - Not keeping all prefixes when there are multiple on a single
    element, which is the case with the Preservica metadata (this is
    dealt with in a few commits)

The above means that parsing XML and then `tostring`ing it using the
`xml` library does some quite unexpected things: the XML does not lookg
the same because prefixes are changed or lost.

It might be possible to workaround the above with registering
namespaces/similar, and perhaps Preservica is a bit flexible with
regards to namespace. However, `lxml` "just works": i.e. if you parse
some XML and then `tostring` it, it is very similar to the output,
including namespaces. Judging this is worth the extra dependency.